refactor: organiza componentes UI e tipos em diretórios modulares
