+++ src/lib/components/ui/card/types.ts
@@ -1,0 +1,8 @@
+// Tipos para Cartão de Métrica
+export interface MetricCardProps {
+	label: string;
+	value: string;
+	change: string;
+	period: string;
+	positive: boolean;
+}
