+++ src/routes/layout.css
@@ -12,7 +12,7 @@
 	--card-foreground: oklch(0.147 0.004 49.25);
 	--popover: #ffffff;
 	--popover-foreground: oklch(0.147 0.004 49.25);
-	--primary: #22c55e;
+	--primary: oklch(71.89% 0.157 132.15);
 	--primary-foreground: #ffffff;
 	--secondary: #e0ede0;
 	--secondary-foreground: oklch(0.21 0.006 285.885);
@@ -23,16 +23,16 @@
 	--destructive: oklch(0.577 0.245 27.325);
 	--border: #e0e0e0;
 	--input: #e0e0e0;
-	--ring: #22c55e;
+	--ring: oklch(71.89% 0.157 132.15);
 	--radius: 0.75rem;
 	--sidebar: #e0ede0;
 	--sidebar-foreground: #1a1a1a;
-	--sidebar-primary: #22c55e;
+	--sidebar-primary: oklch(71.89% 0.157 132.15);
 	--sidebar-primary-foreground: #ffffff;
 	--sidebar-accent: #d0ded0;
 	--sidebar-accent-foreground: #1a1a1a;
 	--sidebar-border: #d0ded0;
-	--sidebar-ring: #22c55e;
+	--sidebar-ring: oklch(71.89% 0.157 132.15);
 	--chart-1: #2563eb;
   	--chart-2: #a855f7;;
 	--color-chrome: #2563eb;
